+++ lib/interval.dart
@@ -1,0 +1,2 @@
+/// Provides a continuous interval over a [Comparable].
+export 'src/interval/interval.dart';