+++ lib/src/interval/extensions.dart
@@ -1,0 +1,3 @@
+import 'interval.dart';
+
+extension NumberIntervalExtension on Interval<num> {}